% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MathFunctions.R
\name{a_n}
\alias{a_n}
\title{Sequence \eqn{a_n} of the Darling-Erdös Law}
\usage{
a_n(n, m)
}
\arguments{
\item{n}{The parameter \eqn{n}}

\item{m}{The parameter \eqn{m}}
}
\value{
The number \eqn{a_n(m)}
}
\description{
Computes \eqn{a_n(m) = \sqrt{b_n(m)/(2 \log \log n)}}, with \eqn{b_n(m)} as
described by \code{\link{b_n}}.
}
\examples{
a_n(5, 2)
}
