% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MathFunctions.R
\name{b_n}
\alias{b_n}
\title{Sequence \eqn{b_n} of the Darling-Erdös Law}
\usage{
b_n(n, m)
}
\arguments{
\item{n}{The parameter \eqn{n}}

\item{m}{The parameter \eqn{m}}
}
\value{
The number \eqn{b_n(m)}
}
\description{
Computes \eqn{b_n(m) = (2 \log \log (n) + (m \log \log \log n) / 2 -
\log(\Gamma(m/2)))^2/(2 \log \log n)}
}
\examples{
b_n(5, 2)
}
