% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MathFunctions.R
\name{ba_kernel_r}
\alias{ba_kernel_r}
\title{Bartlett Kernel}
\usage{
ba_kernel_r(x)
}
\arguments{
\item{x}{Input to \eqn{k(x)}}
}
\value{
Value of kernel
}
\description{
Pure R implementation of the Bartlett kernel.
}
\details{
\deqn{k(x) = \begin{cases} 1 - |x| & |x| \leq 1 \\ 0 & \text{o.w.}
\end{cases}}
}
\examples{
CPAT:::ba_kernel_r(0.5)
}
