% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{banks}
\alias{banks}
\title{Bank Portfolio Returns}
\format{A data frame with 24099 rows and 1 variable:
\describe{
  \item{Banks}{The return of a portfolio representing the banking industry}
}

Row names are dates in YYYY-MM-DD format.}
\source{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
banks
}
\description{
Data set representing the returns of an industry portfolio representing the
banking industry based on company four-digit SIC codes, obtained from the
data library maintained by Kenneth French. Data ranges from July 1, 1926
to October 31, 2017.
}
\keyword{datasets}
