% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{base_file_name}
\alias{base_file_name}
\title{Extract Base File Name}
\usage{
base_file_name(x)
}
\arguments{
\item{x}{String from which to extract base name}
}
\value{
A string containing the base file name without extension
}
\description{
Extract the base name of the file without path or extension.
}
\examples{
CPAT:::base_file_name("~/Documents/test.txt")
}
