% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MathFunctions.R
\name{besselJ_zeros}
\alias{besselJ_zeros}
\title{Compute Zeros of the Bessel Function of the First Kind}
\usage{
besselJ_zeros(b, a = 1, nu = 1)
}
\arguments{
\item{b}{The (one-based) index of the last zero to return}

\item{a}{The (one-based) index of the first zero to return (so \code{a = 1}
represents the first positive zero)}

\item{nu}{The order of the Bessel function}
}
\value{
A vector containing the zeros of the Bessel function
}
\description{
Returns the zeros of the Bessel function of the first kind, \eqn{J_\nu}.
}
\details{
This function is an interface to the function \code{besselJ_zeros_cpp}, a
function written in C++ and serves effectively as an interface to a Boost C++
function \code{cyl_bessel_j_zero}. Thus this function does nothing other than
make the Boost function available to R.

See the references of \code{\link[base]{besselJ}} for more about bessel
functions.
}
\examples{
CPAT:::besselJ_zeros(4)
CPAT:::besselJ_zeros(a = 3, b = 10, nu = 3.5)
}
