% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationUtils.R
\name{bind_power_sim_objs}
\alias{bind_power_sim_objs}
\title{Power Result Data Frame Creation}
\usage{
bind_power_sim_objs(files, crit_value, conv_func, stat_name)
}
\arguments{
\item{files}{A character vector of file names}

\item{crit_value}{The critical value against which to compare a test
statistic}

\item{conv_func}{The function responsible for converting a list containing
simulated statistic values under different conditions to a
\code{data.frame}}

\item{stat_name}{The label of the statistic}
}
\value{
A \code{data.frame} containing power levels
}
\description{
Creates a \code{data.frame} containing power simulation results. Effectively
a better, higher-level interface to \code{\link{power_sim_Zn_to_df}} and
\code{\link{power_sim_Vn_to_df}}.
}
\examples{
\dontrun{
filenames <- c("powerSimulations_sdest_norm_DE.rda",
               "powerSimulations_sdest_ar1_0.5_DE.rda")
bind_power_sim_objs(filenames, crit_value = qdarling_erdos(.95),
                    conv_func = power_sim_Vn_to_df, stat_name = "DE")
}
}
