% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{check_envir_has_objects}
\alias{check_envir_has_objects}
\title{Check An Environment for Objects}
\usage{
check_envir_has_objects(objects, envir = globalenv(),
  blame_string = NULL)
}
\arguments{
\item{objects}{A character vector listing what objects to expect in
\code{envir}}

\item{envir}{The environment to check for objects}

\item{blame_string}{A string that gives more detailed output in error message
if not all files are found; default is the environment
passed to envir}
}
\description{
Check that an environment has expected objects, and stop if it does not.
}
\examples{
x <- 1
CPAT:::check_envir_has_objects(c("x"))
}
