% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StartupMessage.R
\name{CPAT_startup_message}
\alias{CPAT_startup_message}
\title{Create Package Startup Message}
\usage{
CPAT_startup_message()
}
\description{
Makes package startup message.
}
\examples{
CPAT:::CPAT_startup_message()
}
