% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{CXW}
\alias{CXW}
\title{Daily Log Returns of Corrections Corporation of America/CoreCivic Stock (CXW)}
\format{A \code{\link[zoo]{zoo}} time series object with 5109 rows and 1
        variable:
\describe{
  \item{CXW}{The daily (natural) log return of CXW, as a percentage}
}}
\source{
\href{https://www.quandl.com}{Quandl}
}
\usage{
CXW
}
\description{
Data set containing daily log returns (as a percentage) for the stock of
Corrections Corporation of America/CoreCivic, a private prison company, from
July 16, 1997 to October 31, 2017. Stock prices are adjusted for stock splits
and dividends and was obtained from Quandl, via the command
\code{Quandl::Quandl("WIKI/CXW")} (see \code{\link[Quandl]{Quandl}}).
}
\keyword{datasets}
