% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbabilityFunctions.R
\name{dBst}
\alias{dBst}
\title{Density Function of the First Hitting Time of a Bessel Process}
\usage{
dBst(x, b, nu = -1/2, summands = NULL)
}
\arguments{
\item{x}{Points at which to evaluate the density function}

\item{b}{Point in space Bessel process hits}

\item{nu}{The parameter \eqn{\nu > -1} of the Bessel process}

\item{summands}{Number of summands to use in summation; default is to pick
the number of summands with
\code{\link{dBst_summand_solver}} (it could be slow, so for
performance it may be best to pick a fixed number)}
}
\value{
The value of the density function at \code{x}
}
\description{
Density function of the distribution of the first time a Bessel process with
parameter \eqn{\nu > 1} hits \eqn{b > 0}.
}
\details{
Let \eqn{\tau_b^{(\nu)}} be the first time a Bessel process with parameter
\eqn{\nu} hits \eqn{b > 0}. Let \eqn{J_{\nu}(x)} be the Bessel function
(of the first kind) with order \eqn{\nu}, and let \eqn{j_{\nu, k}} be the
\eqn{k}th zero of \eqn{J_{\nu}(x)}. Let \eqn{\Gamma(x)} be the gamma
function. Then the density function of \eqn{\tau_b^{(\nu)}} is

\deqn{\frac{1}{2^\nu b^2 \Gamma(\nu + 1)} \sum_{k = 1}^{\infty} \frac{j_{\nu,
k}^{\nu + 1}}{J_{\nu + 1}(j_{\nu, k})} e^{-\frac{j_{\nu, k}^2}{2b^2}t}}

This was found by differentiating the CDF computed by \code{\link{pBst}}.
}
\examples{
CPAT:::dBst(0.1, 1)
}
