% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MathFunctions.R
\name{dBst_summand_solver}
\alias{dBst_summand_solver}
\title{Find Number of Summands Needed for Numerical Accuracy of \code{dBst}}
\usage{
dBst_summand_solver(x, b, nu = -1/2, error = .Machine$double.eps)
}
\arguments{
\item{x}{Quantile input to PDF}

\item{b}{Point in space Bessel process hits}

\item{nu}{The parameter \eqn{\nu > -1} of the Bessel process}

\item{error}{The desired numerical error of the sum}
}
\value{
Integer for number of summands
}
\description{
Find the number of summands needed to achieve numerical accuracy of the sum
involved in \code{\link{dBst}}.
}
\details{
The number of summands needed is determined by using a loop that runs over
the summands until it encounters a summand that is not greater than the
specified level of numerical accuracy. The index of that last summand is then
returned.
}
\examples{
dBst_summand_solver(1, 1)
}
