% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{dist_conv_plot_tikz}
\alias{dist_conv_plot_tikz}
\title{Create Tikz Plot Demonstrating Rényi-Type Statistic's Convergence in
Distribution}
\usage{
dist_conv_plot_tikz(obj, dist, trim, size, title = "", width = 4,
  height = 3, filename = NULL, makePDF = TRUE, verbose = TRUE)
}
\arguments{
\item{obj}{The list containing the simulations}

\item{dist}{The identifier of the data-generating process that generated the
datasets on which the Rényi-type statistic was computed}

\item{trim}{The identifier of the trimming parameter of the Rényi-type
statistic}

\item{size}{The sample size of the simulated data sets}

\item{title}{The title of the plot}

\item{width}{The width of the plot}

\item{height}{The height of the plot}

\item{filename}{The name of the output file (without extensions; \code{.tex}
and maybe \code{.pdf} files will be created); if \code{NULL},
the name will automatically be determined (of the form
\code{dist_conv_dist_nsize_trim})}

\item{makePDF}{Automatically compile the resulting \code{.tex} file}

\item{verbose}{Print updates about progress (via \code{link[base]{cat}})}
}
\description{
Create a Tikz file containing a plot demonstrating that the Rényi-type
statistic converges in distribution. Optionally, create a PDF as well.
}
\examples{
\dontrun{
ZnSimulations <- list(
  "norm" = list(
    "log" = list(
      "n500" = c(3.206, 1.32, 0.776, 1.262, 0.676)
    )
  )
)
dist_conv_plot_tikz(ZnSimulations, "norm", "log", 500)
}
}
