% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ff}
\alias{ff}
\title{Fama-French Five Factors}
\format{A data frame with 13679 rows and 6 variables:
\describe{
  \item{Mkt.RF}{Market excess returns}
  \item{RF}{The risk-free rate of return}
  \item{SMB}{The return on a diversified portfolio of small stocks minus
             return on a diversified portfolio of big stocks}
  \item{HML}{The return of a portfolio of stocks with a high
             book-to-market (B/M) ratio minus the return of a portfolio
             of stocks with a low B/M ratio}
  \item{RMW}{The return of a portfolio of stocks with robust profitability
             minus a portfolio of stocks with weak profitability}
  \item{CMA}{The return of a portfolio of stocks with conservative
             investment minus the return of a portfolio of stocks with
             aggressive investment}
}

Row names are dates in YYYYMMDD format.}
\source{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
ff
}
\description{
Data set containing the five factors described by
\insertCite{famafrench15;textual}{CPAT}, from the data library maintained by
Kenneth French. Data ranges from July 1, 1963 to October
31, 2017.
}
\keyword{datasets}
