% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{GDPPMI}
\alias{GDPPMI}
\title{GDP vs. PMI}
\format{A \code{\link[zoo]{zoo}} time series object with 285 rows and 2
        variables:
\describe{
  \item{PMI}{Purchasing Manager's Index}
  \item{GDP}{Y.o.Y. GDP}
}}
\source{
\href{https://www.bloomberg.com}{Bloomberg}
}
\usage{
GDPPMI
}
\description{
Year-over-year quarterly U.S. gross domestic product (GDP) and the purchasing
managers' index (PMI) as computed by the Institute for Supply Management,
spanning a period from Q1 1948 to Q1 2019.
}
\keyword{datasets}
