% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{GEO}
\alias{GEO}
\title{Daily Log Returns of GEO Group, Inc. Stock (CXW)}
\format{A \code{\link[zoo]{zoo}} time series object with 5958 rows and 1
        variable:
\describe{
  \item{GEO}{The daily (natural) log return of GEO, as a percentage}
}}
\source{
\href{https://www.quandl.com}{Quandl}
}
\usage{
GEO
}
\description{
Data set containing daily log returns (as a percentage) for the stock of GEO
Group, Inc., a private prison company, from July 29, 1994 to March 27, 2018.
Stock prices are adjusted for stock splits and dividends and was obtained
from Quandl, via the command \code{Quandl::Quandl("WIKI/GEO")} (see
\code{\link[Quandl]{Quandl}}).
}
\keyword{datasets}
