% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationUtils.R
\name{get_expanding_window_pvals}
\alias{get_expanding_window_pvals}
\title{Expanding Window p-Values}
\usage{
get_expanding_window_pvals(dat, m = Inf)
}
\arguments{
\item{dat}{The dataset for which to test for change in mean}

\item{m}{The location of the first potential change point for Andrews' test}
}
\value{
A matrix containing p-values for an expanding sample size, with each
        row corresponding to one observation larger; columns are labeled for
        each statistic
}
\description{
Gets p-values for the CUSUM, Darling-Erdös, Hidalgo-Seo, Andrews, and
Rényi-type tests when applied to an expanding window of data.
}
\examples{
if (require("foreach") & require("doParallel")) {
  get_expanding_window_pvals(rnorm(1000), m = 900)
}
}
