% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationUtils.R
\name{get_expanding_window_pvals_reg}
\alias{get_expanding_window_pvals_reg}
\title{Expanding Window p-Values for Regression Models}
\usage{
get_expanding_window_pvals_reg(formula, data, min_n = 3, m = Inf,
  verbose = FALSE)
}
\arguments{
\item{formula}{The regression model formula, which will be passed to
\code{\link[stats]{lm}}}

\item{data}{A \code{data.frame}, the dataset for which to test for
structural change}

\item{min_n}{An integer; the minimum sample size}

\item{m}{The location of the first potential change point for Andrews' test}

\item{verbose}{If \code{TRUE}, send messages to output}
}
\value{
A matrix containing p-values for an expanding sample size, with each
        row corresponding to one observation larger; columns are labeled for
        each statistic
}
\description{
Gets p-values for the CUSUM, Darling-Erdös, Hidalgo-Seo, Andrews, and
Rényi-type tests when applied to an expanding window of data for a regression
model.
}
\examples{
x <- rnorm(1000)
y <- 1 + 2 * x + rnorm(1000)
df <- data.frame(x, y)
if (require("foreach") & require("doParallel")) {
  get_expanding_window_pvals_reg(y ~ x, data = df, min_n = 4, m = 900)
}
}
