% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangePointTests.R
\name{getLongRunWeights}
\alias{getLongRunWeights}
\title{Weights for Long-Run Variance}
\usage{
getLongRunWeights(n, bandwidth, kernel = "ba")
}
\arguments{
\item{n}{Length of weights' vector}

\item{bandwidth}{A number for the bandwidth}

\item{kernel}{The kernel function; see \code{\link[cointReg]{getLongRunVar}}
for possible values}
}
\value{
List with components \code{w} containing the vector of weights and
        \code{upper}, the index of the largest non-zero entry in \code{w}
}
\description{
Compute some weights for long-run variance. This code comes directly from the
source code of \pkg{cointReg}; see \code{\link[cointReg]{getLongRunWeights}}.
}
\examples{
CPAT:::getLongRunWeights(10, 1)
}
