% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{gold}
\alias{gold}
\title{Daily Gold Price}
\format{A \code{\link[zoo]{zoo}} time series object with 13010 observations}
\source{
\href{https://www.quandl.com}{Quandl}
}
\usage{
gold
}
\description{
Data set containing daily gold prices (per troy ounce). More specifically,
the morning London price fixing (from the London Bullion Market Association),
ranging from 1968 to June 24, 2019. Data was obtained from Quandl, via the
command \code{Quandl::Quandl("LBMA/GOLD")} (see
\code{\link[Quandl]{Quandl}}).
}
\keyword{datasets}
