% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangePointTests.R
\name{HR.test}
\alias{HR.test}
\title{Rényi-Type Test}
\usage{
HR.test(x, formula = NULL, kn = log, use_kernel_var = FALSE,
  stat_plot = FALSE, kernel = "ba", bandwidth = "and")
}
\arguments{
\item{x}{Data to test for change in mean}

\item{formula}{The regression formula, which will be passed to
\code{\link[stats]{lm}}}

\item{kn}{A function corresponding to the trimming parameter \eqn{t_T}; by
default, the square root function}

\item{use_kernel_var}{Set to \code{TRUE} to use kernel methods for long-run
covariance matrix estimation (typically used when the
data is believed to be correlated), and \code{FALSE} to
use an estimator that assumes the data is uncorrelated;
see \code{\link{get_lrv_arr}} for more details}

\item{stat_plot}{Whether to create a plot of the values of the statistic at
all potential change points}

\item{kernel}{If character, the identifier of the kernel function as used in
\pkg{cointReg} (see \code{\link[cointReg]{getLongRunVar}}); if
function, the kernel function to be used for long-run variance
estimation (default is the Bartlett kernel in \pkg{cointReg})}

\item{bandwidth}{If character, the identifier for how to compute the
bandwidth as defined in \pkg{cointReg} (see
\code{\link[cointReg]{getBandwidth}}); if function, a function
to use for computing the bandwidth; if numeric, the bandwidth
value to use (the default is to use Andrews' method, as used in
\pkg{cointReg})}
}
\value{
A \code{htest}-class object containing the results of the test
}
\description{
Performs the (univariate) Rényi-type test for change in mean, as described in
\insertCite{horvathricemiller19}{CPAT} or the Rényi-type test for regression
models, as described in \insertCite{horvathmillerrice20}{CPAT}. This is
effectively an interface to \code{\link{stat_Zn}} and
\code{\link{stat_Zn_reg}}; see their documentation for more details. p-values
are computed using \code{\link{pZn}}, which represents the limiting
distribution of the test statistic under the null hypothesis, which
represents the limiting distribution of the test statistic under the null
hypothesis when \code{kn} represents a sequence \eqn{t_T} satisfying \eqn{t_T
\to \infty} and \eqn{t_T/T \to 0} as \eqn{T \to \infty}.
(\code{\link[base]{log}} and \code{\link[base]{sqrt}} should be good
choices.)
}
\examples{
HR.test(rnorm(1000))
HR.test(rnorm(1000), use_kernel_var = TRUE, kernel = "bo", bandwidth = "nw")
x <- rnorm(1000)
y <- 1 + 2 * x + rnorm(1000)
df <- data.frame(x, y)
HR.test(df, formula = y ~ x, kn = sqrt, use_kernel_var = TRUE, kernel = "qs",
        bandwidth = "and")
}
\references{
\insertAllCited{}
}
