% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{is.formula}
\alias{is.formula}
\title{Check For Formulas}
\usage{
is.formula(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
\code{TRUE} if \code{x} is a \code{\link[stats]{formula}},
        \code{FALSE} otherwise
}
\description{
Checks if an object is a formula.
}
\examples{
CPAT:::is.formula(y ~ x)
CPAT:::is.formula(2)
}
