% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{loop_print_iterator}
\alias{loop_print_iterator}
\title{Command-Line Utility for Printing Through Loop}
\usage{
loop_print_iterator(obj, accessor)
}
\arguments{
\item{obj}{The object through which to iterate}

\item{accessor}{A function with parameters \code{index} and \code{obj} that
will access the entry in position \code{index} of \code{obj}
and return the result}
}
\description{
This is a tool for moving through an iterable object in the R command line.
\code{obj} is the object being iterated through, and \code{accessor} is a
function that can take this object and some index and return the item in that
(integer) index of the object. This is then printed to the screen until
stopped.
}
\details{
When the utility is called, it enters a loop. The user can type the index of
the object she wishes to see, or \code{j} for the previous entry, or \code{k}
for the next entry. Pressing \code{q} ends the loop.
}
\examples{
\dontrun{
loop_print_iterator(1:10, function(index, obj) {obj[[index]]})
}
}
