% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{lrv_plot_tikz}
\alias{lrv_plot_tikz}
\title{Long-Run Variance Estimation Simulations Plot}
\usage{
lrv_plot_tikz(data, n, ker_name, true_lrv, phi = NULL, xrange = NULL,
  width = 4.5, height = 3.5, filename = NULL, verbose = FALSE,
  makePDF = TRUE)
}
\arguments{
\item{data}{A \code{data.frame} containing the data to plot}

\item{n}{The sample size of simulated data sets for which to plot an
estimated distribution}

\item{ker_name}{The name of the kernel function used in the LRV estimator}

\item{true_lrv}{The value of the true long-run variance}

\item{phi}{The autocorrelation parameter of the simulated data sets to plot;
if \code{NULL}, the data is assumed to have been generated with a
GARCH(1,1) process}

\item{xrange}{The limits of the horizontal axis of the plot}

\item{width}{The width of the plot}

\item{height}{The height of the plot}

\item{filename}{The name of the file to save output (without stems; files
with this string appended with \code{.tex} and maybe
\code{.pdf} will be created); if \code{NULL}, a file name
will automatically be chosen (of the form
\code{lrv_est_plot_ker_name_phi})}

\item{verbose}{Print updates about progress (via \code{link[base]{cat}})}

\item{makePDF}{Automatically compile the resulting \code{.tex} file}
}
\description{
Create a Tikz plot of the estimated distribution of LRV estimators
}
\examples{
\dontrun{
plotlist <- data.frame(val = c(0.649, 0.965, 0.905),
                       n   = c(50,    50,    50),
                       phi = c(0,     0,     0))
lrv_plot_tikz(plotlist, 50, ker_name = "bartlett", true_lrv = 1)
}
}
