% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{MktGP}
\alias{MktGP}
\title{Market Returns and the Gold/Platinum Ratio}
\format{A \code{\link[zoo]{zoo}} time series object with 6764 rows and 3
        variables:
\describe{
  \item{Mkt.RF}{The column \code{Mkt.RF} from \code{\link{ff}}}
  \item{GP}{The log ratio of \code{\link{gold}} and \code{\link{platinum}}}
  \item{Mkt.RF.Smooth}{63-day moving average of \code{Mkt.RF}}
}}
\usage{
MktGP
}
\description{
Data set containing excess market returns and the gold/platinum price
ratio, starting from July 3, 1990 and ending October 31, 2017.
}
\keyword{datasets}
