% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{natural_gas_demand}
\alias{natural_gas_demand}
\title{U.S. Natural Gas Consumption, Prices, and GDP}
\format{A time series object with 148 rows and 3 variables:
\describe{
  \item{gdp}{U.S. GDP in millions of dollars, from FRED}
  \item{price}{Average natural gas price in dollars per thousand cubic feet,
               from the EIA}
  \item{consumption}{Natural gas consumption in billions of cubic feet, from
                     the EIA}
}}
\source{
\href{https://fred.stlouisfed.org/series/NA000334Q}{FRED}
        \href{https://www.eia.gov/totalenergy/data/browser/}{EIA}
}
\usage{
natural_gas_demand
}
\description{
Data set containing U.S. total consumption of natural gas and average price
per quarter, along with U.S. GDP. The period spans from Q1 1976 to Q4 2012.
}
\details{
Raw data was obtained from the Federal Reserve Economic Database and from the
U.S. Energy Information Administration.
}
\keyword{datasets}
