% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{OilDollar}
\alias{OilDollar}
\title{Oil Price versus Strength of the Dollar}
\format{A time series object with 402 rows and 2 variables:
\describe{
  \item{Oil}{Crude oil prices: West Texas intermediate}
  \item{DollarExchange}{Real trade weighted U.S. dollar index for broad
                        goods}
}}
\source{
\href{https://fred.stlouisfed.org/series/MCOILWTICO}{FRED: Oil price}
        \href{https://fred.stlouisfed.org/series/TWEXBPA}{FRED: Dollar index}
}
\usage{
OilDollar
}
\description{
Data set containing crude oil prices and an index for the strength of the
U.S. dollar.
}
\details{
Raw data was obtained from the Federal Reserve Economic Database. The data
spans 1986 to June 2019 with monthly frequency.
}
\keyword{datasets}
