% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MathFunctions.R
\name{pa_kernel_r}
\alias{pa_kernel_r}
\title{Parzen Kernel}
\usage{
pa_kernel_r(x)
}
\arguments{
\item{x}{Input to \eqn{k(x)}}
}
\value{
Value of the kernel
}
\description{
Pure R implementation of the Parzen kernel.
}
\details{
\deqn{k(x) = \begin{cases} 1 - 6x^2 + 6|x|^3 & |x| \leq 1/2 \\ 
2(1 - |x|)^3 & 1/2 < |x| \leq 1 \\ 0 & \text{o.w.}
\end{cases}}
}
\examples{
CPAT:::pa_kernel_r(0)
CPAT:::pa_kernel_r(0.4)
CPAT:::pa_kernel_r(0.6)
CPAT:::pa_kernel_r(1.4)
}
