% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbabilityFunctions.R
\name{phidalgo_seo}
\alias{phidalgo_seo}
\title{Hidalgo-Seo Statistic CDF}
\usage{
phidalgo_seo(q)
}
\arguments{
\item{q}{Quantile input to CDF}
}
\value{
If \eqn{Z} is the random variable following the limiting
        distribution, the quantity \eqn{P(Z \leq q)}
}
\description{
CDF of the limiting distribution of the Hidalgo-Seo statistic
}
\examples{
CPAT:::phidalgo_seo(0.1)
}
