% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{platinum}
\alias{platinum}
\title{Daily Platinum Price}
\format{A \code{\link[zoo]{zoo}} time series object with 7382 observations}
\source{
\href{https://www.quandl.com}{Quandl}
}
\usage{
platinum
}
\description{
Data set containing daily platinum prices (per troy ounce). More
specifically, the morning London price fixing (from the London Platinum and
Palladium Society), ranging from April 1990 to June 21, 2019. Data was
obtained from Quandl, via the command \code{Quandl::Quandl("LPPM/PLAT")} (see
\code{\link[Quandl]{Quandl}}).
}
\keyword{datasets}
