% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{power_plot_tikz}
\alias{power_plot_tikz}
\title{Power Curve Plot (By Statistic)}
\usage{
power_plot_tikz(data, d, t, c, s, title = "", legend_pos = "none",
  width = 4.5, height = 3.5, filename = NULL, verbose = FALSE,
  makePDF = TRUE)
}
\arguments{
\item{data}{A \code{data.frame} containing the data to plot}

\item{d}{Label for data-generating process used to simulate the data on which
the statistics were computed}

\item{t}{Label for the trimming parameter of the Rényi-type statistic}

\item{c}{Label for the process that computes the location of change points}

\item{s}{The statistic for which to plot a power curve}

\item{title}{The title of the plot}

\item{legend_pos}{A string to be passed to \code{link[ggplot2]{theme}} (the
\code{legend.position} argument) identifying where to place
the legend}

\item{width}{The width of the plot}

\item{height}{The height of the plot}

\item{filename}{The name of the file to save output (without stems; files
with this string appended with \code{.tex} and maybe
\code{.pdf} will be created); if \code{NULL}, the name will
be automatically determined}

\item{verbose}{Print updates about progress (via \code{link[base]{cat}})}

\item{makePDF}{Automatically compile the resulting \code{.tex} file}
}
\description{
Create a Tikz plot of the power curves of a statistic, with each sample size
having its own curve.
}
\examples{
\dontrun{
pdat <- data.frame(power = c(0.8926, 0.8714, 0.8296, 0.7936),
                   stat  = c("de",   "de",   "de",   "de"),
                   dist  = c("norm", "norm", "norm", "norm"),
                   kn    = c("log",  "log",  "log",  "log"),
                   n     = c(50,     50,     50,     50),
                   cpt   = c("c4rt", "c4rt", "c4rt", "c4rt"),
                   delta = c(-2.0,   -1.9,   -1.8,   -1.7))
power_plot_tikz(pdat, "norm", "log", "c4rt", "de")
}
}
