% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{power_plot_tikz_by_n}
\alias{power_plot_tikz_by_n}
\title{Power Curve Plot}
\usage{
power_plot_tikz_by_n(data, d, t, c, N, statlines, title = "",
  legend_pos = "none", width = 4.5, height = 3.5, filename = NULL,
  verbose = FALSE, makePDF = TRUE)
}
\arguments{
\item{data}{A \code{data.frame} containing the data to plot}

\item{d}{Label for data-generating process used to simulate the data on which
the statistics were computed}

\item{t}{Label for the trimming parameter of the Rényi-type statistic}

\item{c}{Label for the process that computes the location of change points}

\item{N}{The sample size of the simulated data sets on which the statistics
were computed}

\item{statlines}{A character vector where the names of the entries are the
labels of the statistics in the \code{stat} column of
\code{data} and the entries define the line types used by
the \code{values} entry of
\code{\link[ggplot2]{scale_linetype_manual}}}

\item{title}{The title of the plot}

\item{legend_pos}{A string to be passed to \code{link[ggplot2]{theme}} (the
\code{legend.position} argument) identifying where to place
the legend}

\item{width}{The width of the plot}

\item{height}{The height of the plot}

\item{filename}{The name of the file to save output (without stems; files
with this string appended with \code{.tex} and maybe
\code{.pdf} will be created); if \code{NULL}, the name will
be automatically determined}

\item{verbose}{Print updates about progress (via \code{link[base]{cat}})}

\item{makePDF}{Automatically compile the resulting \code{.tex} file}
}
\description{
Create a Tikz plot of the power curves of simulated statistics.
}
\examples{
\dontrun{
pdat <- data.frame(power = c(0.8926, 0.8714, 0.8296, 0.7936),
                   stat  = c("de",   "de",   "de",   "de"),
                   dist  = c("norm", "norm", "norm", "norm"),
                   kn    = c("log",  "log",  "log",  "log"),
                   n     = c(50,     50,     50,     50),
                   cpt   = c("c4rt", "c4rt", "c4rt", "c4rt"),
                   delta = c(-2.0,   -1.9,   -1.8,   -1.7))
power_plot_tikz_by_n(pdat, "norm", "log", "c4rt", 50, c("de" = "solid"))
}
}
