% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationUtils.R
\name{power_sim_stat_df_creator}
\alias{power_sim_stat_df_creator}
\title{Create Power Simulation Results Data Frame}
\usage{
power_sim_stat_df_creator(file_meta, stat_meta, prefix = "",
  alpha = 0.05)
}
\arguments{
\item{file_meta}{The location of a \code{CSV} file that contains file names
and the statistics that those files correspond to}

\item{stat_meta}{The location of a \code{CSV} file that contains statistic
(\code{stat}) labels (used in \code{file_meta}). the name
of the variable for the statistic, and the name of the
function that converts a file (mentioned in
\code{file_meta}) to a \code{data.frame} of power data}

\item{prefix}{Character string representing a prefix for file names mentioned
in \code{file_meta}; could be used for adding path information
to those names, in case the files are not in the working
directory and there is no desire to edit \code{file_meta}'s
data}

\item{alpha}{Numeric for level of significance used in power calculations}
}
\value{
A data frame containing the power simulation data
}
\description{
Creates a \code{data.frame} that contains power simulation results from files
containing power simulations. This function should automate the use of
\code{\link{power_sim_Zn_to_df}} and \code{\link{power_sim_Vn_to_df}} for
collecting power simulation data. It takes two \code{CSV} files, one passed
(as a character string) to \code{file_meta} and the other to
\code{stat_meta}, describing how the files (named and described in
\code{file_meta}) should be handled.
}
\examples{
\dontrun{
power_sim_stat_df_creator("FileStatMeta.csv", "StatMeta.csv")
}
}
