% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationUtils.R
\name{power_sim_Vn_to_df}
\alias{power_sim_Vn_to_df}
\title{Convert CUSUM-Type Statistic Power Simulation Save List to Data Frame}
\usage{
power_sim_Vn_to_df(obj, crit)
}
\arguments{
\item{obj}{A list containing simulated statistic values}

\item{crit}{The critical value determining whether a statistic should lead to
the rejection of the null hypothesis}
}
\value{
A \code{data.frame} summarizing the results of the data stored in
        \code{obj}
}
\description{
This function will convert the power simulation data generated in a list in
our simulation scripts to a \code{data.frame}. Given such a list and a
critical value to determine whether the null hypothesis should be rejected,
the function will return a \code{data.frame} with columns \code{power},
\code{stat}, \code{dist}, \code{n}, \code{cpt}, and \code{delta},
which correspond to: the empirical power of the statistic; the identifier of
the statistic; the generating distribution of the statistic was computed on;
the identifier of how change points were computed; and the size of the
change.
}
\examples{
saveobj <- list("norm" = list(
                  "n50" = list(
                    "c4rt" = list(
                      "d_0" = c(1.551, 1.276, 1.348, 1.982, 1.423)
))))
CPAT:::power_sim_Vn_to_df(saveobj, CPAT:::qkolmogorov(.95))
}
