% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{prod_comp}
\alias{prod_comp}
\title{Productivity and Compensation Indices for the U.S. Economy}
\format{A time series object with 279 rows and 2 variables:
\describe{
  \item{productivity}{Labor productivity (output per hour).}
  \item{compensation}{Real hourly compensation in business sector}
}}
\source{
\href{https://www.bls.gov/}{U.S. Bureau of Labor Statistics}
}
\usage{
prod_comp
}
\description{
Data containing indices for (hourly) productivity and compensation of U.S.
workers from the U.S. Bureau of Labor Statistics. (2009 is 100.)
}
\keyword{datasets}
