% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbabilityFunctions.R
\name{qBst}
\alias{qBst}
\title{Bessel Process First Hitting Time Quantile Function}
\usage{
qBst(p, b, nu = -1/2, summands = NULL, interval = c(0, 100),
  tol = .Machine$double.eps, ...)
}
\arguments{
\item{p}{The probability associated with the desired quantile}

\item{b}{Point in space Bessel process hits}

\item{nu}{The parameter \eqn{\nu > -1} of the Bessel process}

\item{summands}{Number of summands to use in summation; default is to pick
the number of summands with
\code{\link{pBst_summand_solver}} (it could be slow, so for
performance it may be best to pick a fixed number)}

\item{interval, tol, ...}{Arguments to be passed to
\code{\link[stats]{uniroot}}}
}
\value{
The quantile associated with \code{p}
}
\description{
Quantile function of the distribution of the first time a Bessel process with
parameter \eqn{\nu > -1} hits \eqn{b > 0}.
}
\details{
This function uses \code{\link[stats]{uniroot}} for finding this quantity,
and many of the the accepted parameters are arguments for that function; see
its documentation for more details.
}
\examples{
CPAT:::qBst(0.5, b = 1)
}
