% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbabilityFunctions.R
\name{qdarling_erdos}
\alias{qdarling_erdos}
\title{Darling-Erdös Statistic Limiting Distribution Quantile Function}
\usage{
qdarling_erdos(p)
}
\arguments{
\item{p}{The probability associated with the desired quantile}
}
\value{
The quantile associated with \code{p}
}
\description{
Quantile function for the limiting distribution of the Darling-Erdös
statistic.
}
\examples{
CPAT:::qdarling_erdos(0.5)
}
