% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbabilityFunctions.R
\name{qhidalgo_seo}
\alias{qhidalgo_seo}
\title{Hidalgo-Seo Statistic Limiting Distribution Quantile Function}
\usage{
qhidalgo_seo(p)
}
\arguments{
\item{p}{The probability associated with the desired quantile}
}
\value{
A The quantile associated with \code{p}
}
\description{
Quantile function for the limiting distribution of the Hidalgo-Seo statistic
}
\examples{
CPAT:::qhidalgo_seo(0.5)
}
