% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MathFunctions.R
\name{qs_kernel_r}
\alias{qs_kernel_r}
\title{Quadratic Spectral Kernel}
\usage{
qs_kernel_r(x)
}
\arguments{
\item{x}{Input to \eqn{k(x)}}
}
\value{
Value of the kernel
}
\description{
Pure R implementation of the quadratic spectral kernel.
}
\details{
\deqn{k(x) = \frac{25}{12\pi^2 x^2} \left(\frac{\sin(6\pi x/5)}{6\pi x/5} -
\cos(6\pi x / 5)\right)}
}
\examples{
CPAT:::qs_kernel_r(0.4)
}
