% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbabilityFunctions.R
\name{qZn}
\alias{qZn}
\title{Rènyi-Type Statistic Quantile Function}
\usage{
qZn(p, d = 1, summands = 500, interval = c(0, 100),
  tol = .Machine$double.eps, ...)
}
\arguments{
\item{p}{Value of the CDF at the quantile}

\item{summands}{Number of summands for infinite sum}

\item{interval, tol, ...}{Arguments to be passed to
\code{\link[stats]{uniroot}}}
}
\value{
The quantile associated with \code{p}
}
\description{
Quantile function for the limiting distribution of the Rènyi-type statistic.
}
\details{
This function uses \code{\link[stats]{uniroot}} for finding this quantity,
and many of the the accepted parameters are arguments for that function; see
its documentation for more details.
}
\examples{
CPAT:::qZn(0.5)
}
