% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangePointTests.R
\name{stat_hs_reg}
\alias{stat_hs_reg}
\title{Regression Model Hidalgo-Seo Statistic}
\usage{
stat_hs_reg(formula, data, m = sqrt, estimate = FALSE,
  get_all_vals = FALSE)
}
\arguments{
\item{formula}{A \code{link[base]{formula}} that describes the regression
model}

\item{data}{A \code{\link[base]{data.frame}}-like object containing the data
set; should be able to be passed to the \code{data} argument of
\code{\link[stats]{lm}}}

\item{m}{If numeric, the number of terms of the periodogram to sum; if a
function, how to compute the number of terms to sum (will be passed
the number of rows of \code{data})}

\item{estimate}{Set to \code{TRUE} to return the estimated location of the
change point}

\item{get_all_vals}{If \code{TRUE}, return all values for the statistic at
every tested point in the data set}
}
\value{
If both \code{estimate} and \code{get_all_vals} are \code{FALSE}, the
        value of the test statistic; otherwise, a list that contains the test
        statistic and the other values requested (if both are \code{TRUE},
        the test statistic is in the first position and the estimated change
        point in the second)
}
\description{
Compute the Hidalgo-Seo statistic indended for detecting change in linear
models (estimated via least squares regression).
}
\details{
For a data set \eqn{(y_t, x_t)} with \eqn{n} observations, \eqn{y_t \in
\mathbf{R}}, and \eqn{x_t \in \mathbf{R}^d}, the test statistic is

\deqn{\max_{d < s \leq n - d} 	(\mathcal{LM}(s) - B_n)/A_n}

where \eqn{a_n = \sqrt{2 \log \log n}}; \eqn{b_n = a_n^2 + d \log \log \log n
/ 2 - \log \Gamma(d/2)}; \eqn{A_n = b_n/a_n^2}; \eqn{B_n = b_n^2/a_n^2};
\eqn{\hat{\beta}} is the least-squares estimate of the linear regression
model coefficients; \eqn{\hat{u}_t = y_t - \hat{\beta}^T x_t} are the
residuals of the model;

\deqn{\mathcal{LM}(s) = \left(\frac{n}{\hat{\Delta}(\hat{\beta}) s (n - s)}
\right) \left(\sum_{t = 1}^s x_t \hat{u}_t \right)^T\left(\sum_{t = 1}^s x_t
\hat{u}_t \right)^T}

and \eqn{\hat{\Delta}(\hat{\beta})} is the long-run variance estimator

\deqn{\hat{\Delta}(\beta) = \frac{1}{m} \sum_{j = 1}^m I\left( \frac{2\pi
j}{n}; \beta\right)}

where \eqn{I(\cdot ; \beta)} is the periodogram estimated from the residuals
when the regression model coefficients are given by \eqn{\beta}. This is the
test statistic suggested by the procedure introduced in
\insertCite{hidalgoseo13}{CPAT}.

The parameter \eqn{m} described above can be controlled via the function
parameter \code{m}, which can be either numeric or a function that returns
numeric values.
}
\examples{
x <- rnorm(100)
y <- 1 + 2 * x + rnorm(100)
df <- data.frame("x" = x, "y" = y)
CPAT:::stat_hs_reg(y ~ x, data = df)
}
\references{
\insertAllCited{}
}
