% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangePointTests.R
\name{stat_Zn_reg}
\alias{stat_Zn_reg}
\title{Compute the Rényi-Type Statistic for Stability in Linear Regression Models}
\usage{
stat_Zn_reg(formula, data, kn = function(n) {     floor(sqrt(n)) },
  estimate = FALSE, use_kernel_var = FALSE, custom_var = NULL,
  kernel = "ba", bandwidth = "and", get_all_vals = FALSE,
  fast = FALSE)
}
\arguments{
\item{formula}{The regression formula, which will be passed to
\code{\link[stats]{lm}}}

\item{data}{\code{data.frame} containing the data}

\item{kn}{A function corresponding to the trimming parameter \eqn{t_T}; by
default, the square root function}

\item{estimate}{Set to \code{TRUE} to return the estimated location of the
change point}

\item{use_kernel_var}{Set to \code{TRUE} to use kernel methods for long-run
covariance matrix estimation (typically used when the
data is believed to be correlated), and \code{FALSE} to
use an estimator that assumes the data is uncorrelated;
see \code{\link{get_lrv_arr}} for more details}

\item{custom_var}{Either a \eqn{d \times d \times T} array of symmetric
positive-definite matrices containing covariance matrices
or a function producing such an array}

\item{kernel}{If character, the identifier of the kernel function as used in
\pkg{cointReg} (see \code{\link[cointReg]{getLongRunVar}}); if
function, the kernel function to be used for long-run variance
estimation (default is the Bartlett kernel in \pkg{cointReg})}

\item{bandwidth}{If character, the identifier for how to compute the
bandwidth as defined in \pkg{cointReg} (see
\code{\link[cointReg]{getBandwidth}}); if function, a function
to use for computing the bandwidth; if numeric, the bandwidth
value to use (the default is to use Andrews' method, as used in
\pkg{cointReg})}

\item{get_all_vals}{If \code{TRUE}, return all values for the statistic at
every tested point in the data set}

\item{fast}{If \code{TRUE}, the test statistic is computed quickly but at a
potential loss of numerical accuracy (by solving the normal
equations); otherwise, use slower but more numerically stable
solution techniques}
}
\value{
If both \code{estimate} and \code{get_all_vals} are \code{FALSE}, the
        value of the test statistic; otherwise, a list that contains the test
        statistic and the other values requested (if both are \code{TRUE},
        the test statistic is in the first position and the estimated change
        point in the second)
}
\description{
This function computes the Rényi-type statistic for detecting structural
change in linear regression models.
}
\details{
TODO: EXTENDED DESCRIPTION

TODO: THIS FUNCTION DOES NOT WORK AS MARKETED BECAUSE WE'RE STILL WORKING ON
THE THEORY.
}
\examples{
x <- rnorm(1000, mean = 4)
y <- 1 + 2 * x + rnorm(1000)
df <- data.frame(x, y)
CPAT:::stat_Zn_reg(y ~ x, data = df)
}
