% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{stop_with_message}
\alias{stop_with_message}
\title{Check For Condition and Stop With Message}
\usage{
stop_with_message(bool, message = NULL)
}
\arguments{
\item{bool}{Condition to check; if \code{FALSE}, \code{\link[base]{stop}} is
called}

\item{message}{Message to report if \code{\link[base]{stop}} is called}
}
\description{
Check if \code{bool} is \code{TRUE}; if not, stop and report \code{message}
}
\examples{
x <- 1
CPAT:::stop_with_message(x == 1, message = "x is not 1")
}
