% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MathFunctions.R
\name{th_kernel_r}
\alias{th_kernel_r}
\title{Tukey-Hanning Kernel}
\usage{
th_kernel_r(x)
}
\arguments{
\item{x}{Input to \eqn{k(x)}}
}
\value{
Value of the kernel
}
\description{
Pure R implementation of the Tukey-Hanning kernel.
}
\details{
\deqn{k(x) = \begin{cases} (1 + \cos(\pi x))/2 & |x| \leq 1 \\ 0 &
\text{o.w.} \end{cases}}
}
\examples{
CPAT:::th_kernel_r(0.5)
}
