% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MathFunctions.R
\name{tr_kernel_r}
\alias{tr_kernel_r}
\title{Truncated Kernel}
\usage{
tr_kernel_r(x)
}
\arguments{
\item{x}{Input to \eqn{k(x)}}
}
\value{
Value of kernel
}
\description{
Pure R implementation of the truncated kernel.
}
\details{
\deqn{k(x) = \begin{cases} 1 & |x| \leq 1 \\ 0 & \text{o.w.}
\end{cases}}
}
\examples{
CPAT:::tr_kernel_r(0.5)
}
