% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{usa_gbr_price}
\alias{usa_gbr_price}
\title{United States and United Kingdom CPI and Exchange Rates}
\format{A data frame with 578 rows and 3 variables:
\describe{
  \item{USA_CPI}{United States CPI, with 2015 having an index of 100.}
  \item{GBR_CPI}{United Kingdom CPI, with 2015 having an index of 100.}
  \item{GBPUSD}{The exchange rate between the British pound sterling and the
                U.S. dollar, in pounds per dollar.}
}

Row names are month and year.}
\source{
\href{https://fred.stlouisfed.org/series/EXUSUK}{FRED},
        \href{https://data.oecd.org/price/inflation-cpi.htm}{OECD}
}
\usage{
usa_gbr_price
}
\description{
Data set containing the CPI of the United Kingdom and the United States and
the exchange rate between the countries' currencies. Data is at monthly
frequency starting January 1971 and ending February 2019.
}
\details{
CPI data is from the OECD database (downloaded April 5th, 2019) and exchange
rate data is from the FRED database (downloaded April 6th, 2019).
}
\keyword{datasets}
