% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangePointTests.R
\name{wrapped_dynlm}
\alias{wrapped_dynlm}
\title{Dynamic Linear Models and Time Series Regression}
\usage{
wrapped_dynlm(formula, data)
}
\arguments{
\item{formula}{A \code{\link[stats]{formula}} describing the linear model to
be fit}

\item{data}{A \code{\link[base]{data.frame}} or time series object (e.g.,
\code{\link[stats]{ts}} or \code{\link[zoo]{zoo}}), containing
the variables in the model}
}
\value{
The result of the call \code{dynlm(formula, data)}
}
\description{
Interface to \code{\link[stats]{lm.wfit}} for fitting dynamic linear models
and time series regression relationships; a wrapping of the function
\code{\link[dynlm]{dynlm}} from \pkg{dynlm}.
}
\examples{
y <- ts(rnorm(100))
CPAT:::wrapped_dynlm(y ~ L(y), data = y)
}
