% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbabilityFunctions.R
\name{dZn}
\alias{dZn}
\title{Rényi-Type Statistic Limiting Distribution Density Function}
\usage{
dZn(x, summands = 500)
}
\arguments{
\item{x}{Point at which to evaluate the density function (note that this
parameter is not vectorized)}

\item{summands}{Number of summands to use in summation (the default should be
machine accurate)}
}
\value{
Value of the density function at \eqn{x}
}
\description{
Function for computing the value of the density function of the limiting
distribution of the Rényi-type statistic.
}
\examples{
CPAT:::dZn(1)
}
