% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangePointTests.R
\name{get_lrv_arr}
\alias{get_lrv_arr}
\title{Long-Run Covariance Matrix Estimation for Early Change and Block
Heteroskedasticity}
\usage{
get_lrv_arr(X, kernel = "ba", bandwidth = "and")
}
\arguments{
\item{X}{The data matrix}

\item{kernel}{If character, the identifier of the kernel function as used in
\pkg{cointReg} (see \code{\link[cointReg]{getLongRunVar}}); if
function, the kernel function to be used for long-run variance
estimation (default is the Bartlett kernel)}

\item{bandwidth}{If character, the identifier for how to compute the
bandwidth as defined in \pkg{cointReg} (see
\code{\link[cointReg]{getBandwidth}}); if function, a function
to use for computing the bandwidth; if vector, the bandwidth
values to use (the default is to use Andrews' method)}
}
\value{
A 3D array of estimate of the long-run covariance matrices
}
\description{
Estimates the covariance matrix of the process \eqn{\{x_t \epsilon_t\}} in a
block-heteroskedasticity context; see \insertCite{horvathmillerrice20}{CPAT}
for more details.
}
\details{
While the documentation refers to \pkg{cointReg}, that package is never used
here. The kernels and methods referred to have implementations specific to
this context. This is due to the computational demand of computing
\code{nrow(X)} separate (but related) covariance matrices using kernel
methods. Thus this function, by default, uses \pkg{cointReg} notation only,
not its methods.
}
\examples{
x <- rnorm(50)
y <- 1 + 2 * x + rnorm(50)
df <- data.frame(x, y)
X <- model.matrix(y ~ x, data = df)
CPAT:::get_lrv_arr(X)
}
\references{
\insertAllCited{}
}
