% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbabilityFunctions.R
\name{pBst}
\alias{pBst}
\title{CDF of First Hitting Time of Bessel Process}
\usage{
pBst(q, b, nu = -1/2, summands = 500)
}
\arguments{
\item{q}{Quantile input to CDF}

\item{b}{Point in space Bessel process hits}

\item{nu}{The parameter \eqn{\nu > -1} of the Bessel process}

\item{summands}{Number of summands to use in summation}
}
\value{
If \eqn{T} is the random variable as described, \eqn{P(T \leq q)}
}
\description{
CDF of the distribution of the first time a Bessel process with parameter
\eqn{\nu > -1} hits \eqn{b > 0}.
}
\examples{
CPAT:::pBst(1, 1)
}
