% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbabilityFunctions.R
\name{pZn}
\alias{pZn}
\title{Rènyi-Type Statistic CDF}
\usage{
pZn(q, summands = NULL)
}
\arguments{
\item{q}{Quantile input to CDF}

\item{summands}{Number of summands for infinite sum; if \code{NULL},
automatically determined}
}
\value{
If \eqn{Z} is the random variable following the limiting
        distribution, the quantity \eqn{P(Z \leq q)}
}
\description{
CDF for the limiting distribution of the Rènyi-type statistic.
}
\examples{
CPAT:::pZn(0.1)
}
