% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbabilityFunctions.R
\name{pZn}
\alias{pZn}
\title{Rènyi-Type Statistic CDF}
\usage{
pZn(q, d = 1, summands = NULL)
}
\arguments{
\item{q}{Quantile input to CDF}

\item{d}{Dimension parameter}

\item{summands}{Number of summands for infinite sum; if \code{NULL},
automatically determined using
\code{\link{pBst_summand_solver}} (which isn't necessarily
fast, so consider picking a fixed number if speed is
important)}
}
\value{
If \eqn{Z} is the random variable following the limiting
        distribution, the quantity \eqn{P(Z \leq q)}
}
\description{
CDF for the limiting distribution of the Rènyi-type statistic.
}
\examples{
CPAT:::pZn(0.1)
}
