% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangePointTests.R
\name{stat_Zn_reg}
\alias{stat_Zn_reg}
\title{Compute the Rényi-Type Statistic for Stability in Linear Regression Models}
\usage{
stat_Zn_reg(formula, data, kn = function(n) {     floor(sqrt(n)) },
  estimate = FALSE, use_kernel_var = FALSE, custom_var = NULL,
  kernel = "ba", bandwidth = "and", get_all_vals = FALSE,
  fast = FALSE)
}
\arguments{
\item{formula}{The regression formula, which will be passed to
\code{\link[dynlm]{dynlm}}}

\item{data}{\code{data.frame} containing the data}

\item{kn}{A function corresponding to the trimming parameter \eqn{t_T}; by
default, the square root function}

\item{estimate}{Set to \code{TRUE} to return the estimated location of the
change point}

\item{use_kernel_var}{Set to \code{TRUE} to use kernel methods for long-run
covariance matrix estimation (typically used when the
data is believed to be correlated), and \code{FALSE} to
use an estimator that assumes the data is uncorrelated;
see \code{\link{get_lrv_arr}} for more details}

\item{custom_var}{Either a \eqn{d \times d \times T} array of symmetric
positive-definite matrices containing covariance matrices
or a function producing such an array and having a
parameter \code{x} for the data; if set,
\code{use_kernel_var} is ignored}

\item{kernel}{If character, the identifier of the kernel function as used in
\pkg{cointReg} (see \code{\link[cointReg]{getLongRunVar}}); if
function, the kernel function to be used for long-run variance
estimation (default is the Bartlett kernel in \pkg{cointReg})}

\item{bandwidth}{If character, the identifier for how to compute the
bandwidth as defined in \pkg{cointReg} (see
\code{\link[cointReg]{getBandwidth}}); if function, a function
to use for computing the bandwidth; if numeric, the bandwidth
value to use (the default is to use Andrews' method, as used in
\pkg{cointReg})}

\item{get_all_vals}{If \code{TRUE}, return all values for the statistic at
every tested point in the data set}

\item{fast}{If \code{TRUE}, the test statistic is computed quickly but at a
potential loss of numerical accuracy (by solving the normal
equations); otherwise, use slower but more numerically stable
solution techniques}
}
\value{
If both \code{estimate} and \code{get_all_vals} are \code{FALSE}, the
        value of the test statistic; otherwise, a list that contains the test
        statistic and the other values requested (if both are \code{TRUE},
        the test statistic is in the first position and the estimated change
        point in the second)
}
\description{
This function computes the Rényi-type statistic for detecting structural
change in linear regression models. See
\insertCite{horvathmillerrice20}{CPAT} for more details.
}
\details{
Let \eqn{y_t = x_t^T \beta_t + \epsilon_t}. The definition of the statistic
is

\deqn{\sqrt{t_T} \max_{t_T \leq t \leq T} \sqrt{(\hat{\beta_{t,1} -
\hat{\beta_{t,2}}})^T \hat{D}_{T,t}^{-1} (\hat{\beta}_{1,t} -
\hat{\beta}_{2,t})}}

where \eqn{t_T} is the trimming parameter, \eqn{\hat{\beta}_{1,t}} is the
least-squares model fit for the first half of the data when split at \eqn{t},
\eqn{\hat{\beta}_{2,t}} the same but for the upper half, and

\deqn{\hat{D}_{t,T} = I_{t \leq T/2} (t^{-1} X_{t,1}^T X_{t,1})
\hat{Q_{t,T}^{-1}} (t^{-1} X_{t,1}^T X_{t,1}) + I_{t > T/2} ((T - t)^{-1}
X_{t,2}^T X_{t,2}) \hat{Q_{t,T}^{-1}} ((T - t)^{-1} X_{t,2}^T X_{t,2})}

with \eqn{X_{t,1}} being the data matrix containing the data up to \eqn{t},
\eqn{X_{t,2}} the data matrix containing the data after \eqn{t}, \eqn{I} the
indicator function, and \eqn{Q_{t,T}} a matrix for long-run variance
estimation of the process \eqn{\{x_t \epsilon_t\}}.

Let \eqn{\hat{e}_t = x_t \hat{\epsilon}_t}. When \code{use_kernel_var} is
\code{FALSE}, then the internal estimator for \eqn{Q} is used; it is

\deqn{\hat{Q}_{t,T} = I_{t \leq T/2} \hat{Q}_{1,t} + I_{t > T/2}
\hat{Q}_{t,2}}

with \eqn{\hat{Q}_{1,t} = t^{-1} \sum_{s = 1}^t \hat{e}_s \hat{e}_s^T} and
\eqn{\hat{Q}_{t,2} = (T - t)^{-1} \sum_{s = t + 1}^{T} \hat{e}_s
\hat{e}_s^T}. Otherwise, \code{\link{get_lrv_arr}} is called to estimate this
matrix using kernel methods (unless \code{custom_var} is set, which will then
be used for estimation instead).
}
\examples{
x <- rnorm(1000, mean = 4)
y <- 1 + 2 * x + rnorm(1000)
df <- data.frame(x, y)
CPAT:::stat_Zn_reg(y ~ x, data = df)
}
\references{
\insertAllCited{}
}
